requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==25.1.0
aiohttp==3.13.2
python-interface==1.6.1
injector==0.23.0